Improved getPoints, and added template for main method.
I changed getPoint into a method that will actually pull multiple points out in a single line. This removes the need for a for loop within the main method, and actually tightens down the code found within getPoints as well.

I used the spec layed out in our UML diagram to prototype the main method, though obviously this will need to be changed to meet the functions as written by Jackson, once I can see them.